Add files via upload
Data were cleaned, and formatted and a descriptive analysis of data was done and visualised to understand the distribution of the data. the issue of class imbalance was handled by creating a sample data set to resolve the class imbalance problem with data distribution. The Time series function was used to extract the hour of the transaction from the Unix time column. The correlation of the columns was also obtained to have an idea of the level of correlation amongst various selected columns. Data has been prepared for machine learning modelling.